# tools/usage-dashboard/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CLIProxyAPI Usage Dashboard",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/aliases": {
      "get": {
        "summary": "List Aliases",
        "description": "Return all key aliases.",
        "operationId": "list_aliases_api_v1_aliases_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "summary": "Upsert Alias",
        "description": "Create or update a key alias.",
        "operationId": "upsert_alias_api_v1_aliases_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AliasBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aliases/{account_hash}": {
      "delete": {
        "summary": "Delete Alias",
        "description": "Delete a key alias by account_hash.",
        "operationId": "delete_alias_api_v1_aliases__account_hash__delete",
        "parameters": [
          {
            "name": "account_hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Hash"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AliasBody": {
        "properties": {
          "account_hash": {
            "type": "string",
            "title": "Account Hash"
          },
          "alias": {
            "type": "string",
            "title": "Alias"
          }
        },
        "type": "object",
        "required": [
          "account_hash",
          "alias"
        ],
        "title": "AliasBody"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}